backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
stripe
gunicorn
python-decouple
psycopg2-binary
django-environ


